# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Business Management System API
  version: 1.0.0
paths:
  /customers:
    post:
      summary: Create Customer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created successfully
    get:
      summary: List Customers
      responses:
        '200':
          description: List of customers
  /customers/{dni}:
    get:
      summary: Read Customer
      parameters:
        - in: path
          name: dni
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer details
        '404':
          description: Customer not found
    put:
      summary: Update Customer
      parameters:
        - in: path
          name: dni
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated successfully
        '404':
          description: Customer not found
    delete:
      summary: Delete Customer
      parameters:
        - in: path
          name: dni
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
...
components:
  schemas:
    Customer:
      type: object
      properties:
        nombre:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        dni:
          type: string
      required:
        - nombre
        - telefono
        - direccion
        - dni
...